url-py-service/requirements.txt
flask
flask-cors
redis
gevent
psycopg2-binary
